requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.2
altgraph==0.17.4
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
bleach==6.0.0
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.7
docutils==0.20.1
fonttools==4.43.1
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.40
gobject==0.1.0
greenlet==3.0.1
gst==0.1.0
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
keyring==23.13.1
kiwisolver==1.4.5
langchain==0.0.175
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
mutagen==1.46.0
mypy==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.1
pefile==2023.2.7
Pillow==9.5.0
pkginfo==1.9.6
protobuf==4.25.0
pyarrow==14.0.1
pycryptodomex==3.18.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0
readme-renderer==37.3
referencing==0.31.0
regex==2023.5.5
requests==2.30.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.15
streamlit==1.28.2
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
twine==4.0.2
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.2
validators==0.22.0
watchdog==3.0.0
webencodings==0.5.1
websockets==11.0.3
xlrd==2.0.1
yarl==1.9.2
youtube-summary==0.4
youtube-transcript-api==0.6.0
yt-dlp==2023.3.4
zipp==3.15.0
